Refactor: modular web rather than class
