+++ src/tools/Got.ts
@@ -1,5 +1,5 @@
 import got from 'got';
 
 export default got.extend({
-	headers: { 'User-Agent': 'Melonbot github.com/joachimflottorp/melonbot' },
+	headers: { 'User-Agent': 'Melonbot' },
 });
