+++ src/web/middlewares/logger.ts
@@ -1,0 +1,10 @@
+import express from 'express';
+
+export default (
+	req: express.Request,
+	res: express.Response,
+	next: express.NextFunction,
+) => {
+	console.log(req.url);
+	next();
+};
