--- src/web/routes/common.routes.config.ts
@@ -1,18 +1,0 @@
-import express from 'express';
-
-export abstract class CommonRoutesConfig {
-	app: express.Application;
-	name: string;
-
-	constructor(app: express.Application, name: string) {
-		this.app = app;
-		this.name = name;
-		this.configureRoutes();
-	}
-
-	getName() {
-		return this.name;
-	}
-
-	abstract configureRoutes(): express.Application;
-}
