+++ src/web/views/index.pug
@@ -24,6 +24,6 @@ html
 
                     
                     div(class="dropdown-menu" aria-labelledby="dropdownMenuButton")
-                        input(class="dropdown-item" type="button" onclick="location.href='/bot/login'" value="Login Twitch")
+                        input(class="dropdown-item" type="button" onclick="location.href='/login'" value="Login Twitch")
 
                         input(class="dropdown-item" type="button" onclick="location.href='/bot/commands'" value="View Commands")
