+++ src/web/views/trStats.pug
@@ -11,7 +11,7 @@ script(type="text/javascript" defer).
             commandsHandled: 0
         }
         
-        await fetch("/v1/stats", {
+        await fetch("/api/v1/stats", {
             method: 'GET',
             headers: {
                 'Content-Type': 'application/json'
